+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Unreleased
 Added
 ~~~~~
 
+* finatra-kafka: Expose endOffsets() in FinagleKafkaConsumer. ``PHAB_ID=D263573``
+
 * finatra-kafka-streams: Adding missing ScalaDocs. Adding metric for elapsed state
   restore time. RocksDB configuration now contains a flag for adjusting the number
   of cache shard bits, `rocksdb.block.cache.shard.bits`. ``PHAB_ID=D255771``